backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==43.0.0
Django==4.2.14
django-appconf==1.0.6
django-cors-headers==4.4.0
django-cryptography==1.1
djangorestframework==3.15.2
pycparser==2.22
sqlparse==0.5.1
